@@ -21,7 +21,7 @@ abstract class Base implements Loader {
 	 *
 	 * @var \Required\Traduttore\Repository Repository object.
 	 */
-	protected $repository;
+	protected Repository $repository;
 
 	/**
 	 * Class constructor.
@@ -46,7 +46,7 @@ abstract class Base implements Loader {
 			'%1$straduttore-%2$s-%3$s',
 			get_temp_dir(),
 			$this->repository->get_host(),
-			sanitize_title( $this->repository->get_name() )
+			sanitize_title( $this->repository->get_name() ?? '' )
 		);
 	}
 }
