@@ -1,29 +1,23 @@
 <?php
 /**
  * Class Export
- *
- * @package Traduttore\Tests
  */
 
 namespace Required\Traduttore\Tests;
 
-use GP_Translation_Set;
 use PO;
-use \Required\Traduttore\Export as E;
+use Required\Traduttore\Export as E;
 
 /**
  * Test cases for \Required\Traduttore\Export.
  */
 class Export extends TestCase {
-	/**
-	 * @var GP_Translation_Set
-	 */
-	protected $translation_set;
+	protected \GP_Translation_Set $translation_set;
 
 	public function setUp(): void {
 		parent::setUp();
 
-		$locale = $this->factory->locale->create(
+		$locale = $this->factory()->locale->create(
 			[
 				'english_name' => 'German',
 				'native_name'  => 'Deutsch',
@@ -32,7 +26,7 @@ class Export extends TestCase {
 			]
 		);
 
-		$this->translation_set = $this->factory->translation_set->create_with_project(
+		$this->translation_set = $this->factory()->translation_set->create_with_project(
 			[
 				'locale' => $locale->slug,
 			],
@@ -49,14 +43,14 @@ class Export extends TestCase {
 	}
 
 	public function test_creates_only_po_and_mo_files(): void {
-		$original = $this->factory->original->create(
+		$original = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => 'my-plugin.php',
 			]
 		);
 
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original->id,
 				'translation_set_id' => $this->translation_set->id,
@@ -68,12 +62,15 @@ class Export extends TestCase {
 
 		$actual = $export->export_strings();
 
+		$this->assertIsArray( $actual );
+
 		array_map( 'unlink', $actual );
 
 		$this->assertEqualSets(
 			[
 				'foo-project-de_DE.po',
 				'foo-project-de_DE.mo',
+				'foo-project-de_DE.l10n.php',
 			],
 			array_keys( $actual )
 		);
@@ -84,7 +81,7 @@ class Export extends TestCase {
 		$filename_2 = 'my-super-minified-script';
 
 		/* @var \GP_Original $original_1 */
-		$original_1 = $this->factory->original->create(
+		$original_1 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => $filename_1 . '.js',
@@ -92,14 +89,14 @@ class Export extends TestCase {
 		);
 
 		/* @var \GP_Original $original_2 */
-		$original_2 = $this->factory->original->create(
+		$original_2 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => $filename_2 . '.min.js',
 			]
 		);
 
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_1->id,
 				'translation_set_id' => $this->translation_set->id,
@@ -107,7 +104,7 @@ class Export extends TestCase {
 			]
 		);
 
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_2->id,
 				'translation_set_id' => $this->translation_set->id,
@@ -119,6 +116,8 @@ class Export extends TestCase {
 
 		$actual = $export->export_strings();
 
+		$this->assertIsArray( $actual );
+
 		$json_filename_1 = 'foo-project-de_DE-' . md5( $filename_1 . '.js' ) . '.json';
 		$json_filename_2 = 'foo-project-de_DE-' . md5( $filename_2 . '.js' ) . '.json';
 
@@ -127,12 +126,15 @@ class Export extends TestCase {
 
 		array_map( 'unlink', $actual );
 
+		$this->assertIsString( $json_1 );
+		$this->assertIsString( $json_2 );
 		$this->assertJson( $json_1 );
 		$this->assertJson( $json_2 );
 		$this->assertEqualSets(
 			[
 				'foo-project-de_DE.po',
 				'foo-project-de_DE.mo',
+				'foo-project-de_DE.l10n.php',
 				$json_filename_1,
 				$json_filename_2,
 			],
@@ -143,9 +145,8 @@ class Export extends TestCase {
 	/**
 	 * Modify the mapping of sources to translation entries.
 	 *
-	 * @param array $mapping The mapping of sources to translation entries.
-	 *
-	 * @return array The maybe modified mapping.
+	 * @param array<string, \Translation_Entry[]> $mapping The mapping of sources to translation entries.
+	 * @return array<string, \Translation_Entry[]> The maybe modified mapping.
 	 */
 	public function filter_map_entries_to_source( array $mapping ): array {
 		$mapping['build.js'] = array_merge( $mapping['my-super-script.js'], $mapping['my-other-script.js'] );
@@ -161,7 +162,7 @@ class Export extends TestCase {
 		$filename_target = 'build.js';
 
 		/* @var \GP_Original $original_1 */
-		$original_1 = $this->factory->original->create(
+		$original_1 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => $filename_1,
@@ -169,14 +170,14 @@ class Export extends TestCase {
 		);
 
 		/* @var \GP_Original $original_2 */
-		$original_2 = $this->factory->original->create(
+		$original_2 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => $filename_2,
 			]
 		);
 
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_1->id,
 				'translation_set_id' => $this->translation_set->id,
@@ -184,7 +185,7 @@ class Export extends TestCase {
 			]
 		);
 
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_2->id,
 				'translation_set_id' => $this->translation_set->id,
@@ -197,6 +198,8 @@ class Export extends TestCase {
 		add_filter( 'traduttore.map_entries_to_source', [ $this, 'filter_map_entries_to_source' ] );
 
 		$actual = $export->export_strings();
+
+		$this->assertIsArray( $actual );
 
 		remove_filter( 'traduttore.map_entries_to_source', [ $this, 'filter_map_entries_to_source' ] );
 
@@ -211,11 +214,13 @@ class Export extends TestCase {
 
 		array_map( 'unlink', $actual );
 
+		$this->assertIsString( $json );
 		$this->assertJson( $json );
 		$this->assertEqualSets(
 			[
 				'foo-project-de_DE.po',
 				'foo-project-de_DE.mo',
+				'foo-project-de_DE.l10n.php',
 				$json_filename_target,
 			],
 			array_keys( $actual )
@@ -227,7 +232,7 @@ class Export extends TestCase {
 		$filename_2 = 'my-super-minified-script';
 
 		/* @var \GP_Original $original_1 */
-		$original_1 = $this->factory->original->create(
+		$original_1 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => $filename_1 . '.js',
@@ -235,7 +240,7 @@ class Export extends TestCase {
 		);
 
 		/* @var \GP_Original $original_2 */
-		$original_2 = $this->factory->original->create(
+		$original_2 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => $filename_2 . '.min.js',
@@ -243,28 +248,28 @@ class Export extends TestCase {
 		);
 
 		/* @var \GP_Original $original_3 */
-		$original_3 = $this->factory->original->create(
+		$original_3 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => 'foo.php',
 			]
 		);
 
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_1->id,
 				'translation_set_id' => $this->translation_set->id,
 				'status'             => 'current',
 			]
 		);
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_2->id,
 				'translation_set_id' => $this->translation_set->id,
 				'status'             => 'current',
 			]
 		);
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_3->id,
 				'translation_set_id' => $this->translation_set->id,
@@ -276,14 +281,19 @@ class Export extends TestCase {
 
 		$actual = $export->export_strings();
 
+		$this->assertIsArray( $actual );
+
 		$translations = new PO();
 		$translations->import_from_file( $actual['foo-project-de_DE.po'] );
 
 		$json_filename_1 = 'foo-project-de_DE-' . md5( $filename_1 . '.js' ) . '.json';
 		$json_filename_2 = 'foo-project-de_DE-' . md5( $filename_2 . '.js' ) . '.json';
 
-		$json_1 = json_decode( file_get_contents( $actual[ $json_filename_1 ] ), true );
-		$json_2 = json_decode( file_get_contents( $actual[ $json_filename_2 ] ), true );
+		$json_1 = json_decode( (string) file_get_contents( $actual[ $json_filename_1 ] ), true );
+		$json_2 = json_decode( (string) file_get_contents( $actual[ $json_filename_2 ] ), true );
+
+		$this->assertIsArray( $json_1 );
+		$this->assertIsArray( $json_2 );
 
 		array_map( 'unlink', $actual );
 
@@ -301,7 +311,7 @@ class Export extends TestCase {
 		$filename_3 = 'dist/build.js';
 
 		/* @var \GP_Original $original_1 */
-		$original_1 = $this->factory->original->create(
+		$original_1 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => "$filename_1 $filename_3",
@@ -309,7 +319,7 @@ class Export extends TestCase {
 		);
 
 		/* @var \GP_Original $original_2 */
-		$original_2 = $this->factory->original->create(
+		$original_2 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => "$filename_2 $filename_3",
@@ -317,28 +327,28 @@ class Export extends TestCase {
 		);
 
 		/* @var \GP_Original $original_3 */
-		$original_3 = $this->factory->original->create(
+		$original_3 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => $filename_3,
 			]
 		);
 
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_1->id,
 				'translation_set_id' => $this->translation_set->id,
 				'status'             => 'current',
 			]
 		);
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_2->id,
 				'translation_set_id' => $this->translation_set->id,
 				'status'             => 'current',
 			]
 		);
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_3->id,
 				'translation_set_id' => $this->translation_set->id,
@@ -350,13 +360,15 @@ class Export extends TestCase {
 
 		$actual = $export->export_strings();
 
+		$this->assertIsArray( $actual );
+
 		$json_filename = 'foo-project-de_DE-' . md5( $filename_3 ) . '.json';
 
-		$json = json_decode( file_get_contents( $actual[ $json_filename ] ), true );
+		$json = json_decode( (string) file_get_contents( $actual[ $json_filename ] ), true );
 
 		array_map( 'unlink', $actual );
 
-		$this->assertInternalType( 'array', $json );
+		$this->assertIsArray( $json );
 		$this->assertCount( 4, $json['locale_data']['messages'] );
 		$this->assertArrayHasKey( $original_1->singular, $json['locale_data']['messages'] );
 		$this->assertArrayHasKey( $original_2->singular, $json['locale_data']['messages'] );
@@ -365,6 +377,7 @@ class Export extends TestCase {
 			[
 				'foo-project-de_DE.po',
 				'foo-project-de_DE.mo',
+				'foo-project-de_DE.l10n.php',
 				$json_filename,
 			],
 			array_keys( $actual )
@@ -375,14 +388,14 @@ class Export extends TestCase {
 		$filename_1 = 'my-super-script';
 
 		/* @var \GP_Original $original_1 */
-		$original_1 = $this->factory->original->create(
+		$original_1 = $this->factory()->original->create(
 			[
 				'project_id' => $this->translation_set->project_id,
 				'references' => $filename_1 . '.js',
 			]
 		);
 
-		$this->factory->translation->create(
+		$this->factory()->translation->create(
 			[
 				'original_id'        => $original_1->id,
 				'translation_set_id' => $this->translation_set->id,
@@ -394,15 +407,18 @@ class Export extends TestCase {
 
 		$actual = $export->export_strings();
 
+		$this->assertIsArray( $actual );
+
 		$json_filename_1 = 'foo-project-de_DE-' . md5( $filename_1 . '.js' ) . '.json';
 
 		$json_1 = file_get_contents( $actual[ $json_filename_1 ] );
 
 		array_map( 'unlink', $actual );
 
+		$this->assertIsString( $json_1 );
 		$this->assertJson( $json_1 );
 
-		$json1_encoded = json_decode( $json_1 );
+		$json1_encoded = (object) json_decode( $json_1 );
 		$this->assertSame( $filename_1 . '.js', $json1_encoded->comment->reference );
 	}
 }
