@@ -1,15 +1,12 @@
 <?php
 /**
  * Class Mercurial
- *
- * @package Traduttore\Tests
  */
 
 namespace Required\Traduttore\Tests\Loader;
 
-use \GP_UnitTestCase;
-use \Required\Traduttore\Loader\Mercurial as MercurialLoader;
-use \Required\Traduttore\Project;
+use Required\Traduttore\Loader\Mercurial as MercurialLoader;
+use Required\Traduttore\Project;
 use Required\Traduttore\Repository\Bitbucket;
 use Required\Traduttore\Tests\TestCase;
 
@@ -19,16 +16,13 @@ use Required\Traduttore\Tests\TestCase;
  * @todo Mock shell execution
  */
 class Mercurial extends TestCase {
-	/**
-	 * @var Project
-	 */
-	protected $project;
+	protected Project $project;
 
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$this->project = new Project(
-			$this->factory->project->create(
+			$this->factory()->project->create(
 				[
 					'name'                => 'Sample Project',
 					'slug'                => 'sample-project',
@@ -45,8 +39,9 @@ class Mercurial extends TestCase {
 	}
 
 	public function test_download_repository(): void {
-		$this->markTestSkipped( 'Need to mock shell command execution' );
+		$this->markTestIncomplete( 'Need to mock shell command execution' );
 
+		// @phpstan-ignore-next-line
 		$loader = new MercurialLoader( new Bitbucket( $this->project ) );
 
 		add_filter( 'traduttore.hg_clone_use_https', '__return_true' );
@@ -57,8 +52,9 @@ class Mercurial extends TestCase {
 	}
 
 	public function test_download_existing_repository(): void {
-		$this->markTestSkipped( 'Need to mock shell command execution' );
+		$this->markTestIncomplete( 'Need to mock shell command execution' );
 
+		// @phpstan-ignore-next-line
 		$loader = new MercurialLoader( new Bitbucket( $this->project ) );
 
 		add_filter( 'traduttore.hg_clone_use_https', '__return_true' );
