@@ -1,8 +1,6 @@
 <?php
 /**
- * Class Bitbucket
- *
- * @package Traduttore\Tests\WebhookHandler
+ * Class Bitbucket\WebhookHandler
  */
 
 namespace Required\Traduttore\Tests\WebhookHandler;
@@ -10,22 +8,19 @@ namespace Required\Traduttore\Tests\WebhookHandler;
 use Required\Traduttore\Project;
 use Required\Traduttore\Repository;
 use Required\Traduttore\Tests\TestCase;
-use \WP_REST_Request;
+use WP_REST_Request;
 
 /**
  * Test cases for \Required\Traduttore\WebhookHandler\Bitbucket.
  */
 class Bitbucket extends TestCase {
-	/**
-	 * @var Project
-	 */
-	protected $project;
+	protected Project $project;
 
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$this->project = new Project(
-			$this->factory->project->create(
+			$this->factory()->project->create(
 				[
 					'name'                => 'Sample Project',
 					'source_url_template' => 'https://bitbucket.org/wearerequired/traduttore/blob/master/%file%#L%line%',
@@ -51,10 +46,11 @@ class Bitbucket extends TestCase {
 
 	public function test_invalid_signature(): void {
 		$request = new WP_REST_Request( 'POST', '/traduttore/v1/incoming-webhook' );
-		$request->set_body_params( [] );
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body( (string) wp_json_encode( [] ) );
 		$signature = 'sha256=' . hash_hmac( 'sha256', $request->get_body(), 'foo' );
 		$request->add_header( 'x-event-key', 'repo:push' );
-		$request->add_header( 'x-hub-signature', $signature );
+		$request->add_header( 'x-hub-signature-256', $signature );
 		$response = rest_get_server()->dispatch( $request );
 
 		$this->assertErrorResponse( 'rest_forbidden', $response, 401 );
@@ -62,20 +58,23 @@ class Bitbucket extends TestCase {
 
 	public function test_missing_signature_is_valid(): void {
 		$request = new WP_REST_Request( 'POST', '/traduttore/v1/incoming-webhook' );
-		$request->set_body_params(
-			[
-				'ref'        => 'refs/heads/master',
-				'repository' => [
-					'links'      => [
-						'html' => [
-							'href' => 'https://bitbucket.org/wearerequired/not-traduttore',
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body(
+			(string) wp_json_encode(
+				[
+					'ref'        => 'refs/heads/master',
+					'repository' => [
+						'links'      => [
+							'html' => [
+								'href' => 'https://bitbucket.org/wearerequired/not-traduttore',
+							],
 						],
+						'full_name'  => 'wearerequired/not-traduttore',
+						'scm'        => 'git',
+						'is_private' => false,
 					],
-					'full_name'  => 'wearerequired/not-traduttore',
-					'scm'        => 'git',
-					'is_private' => false,
-				],
-			]
+				]
+			)
 		);
 		$request->add_header( 'x-event-key', 'repo:push' );
 		$response = rest_get_server()->dispatch( $request );
@@ -85,24 +84,27 @@ class Bitbucket extends TestCase {
 
 	public function test_invalid_project(): void {
 		$request = new WP_REST_Request( 'POST', '/traduttore/v1/incoming-webhook' );
-		$request->set_body_params(
-			[
-				'ref'        => 'refs/heads/master',
-				'repository' => [
-					'links'      => [
-						'html' => [
-							'href' => 'https://bitbucket.org/wearerequired/not-traduttore',
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body(
+			(string) wp_json_encode(
+				[
+					'ref'        => 'refs/heads/master',
+					'repository' => [
+						'links'      => [
+							'html' => [
+								'href' => 'https://bitbucket.org/wearerequired/not-traduttore',
+							],
 						],
+						'full_name'  => 'wearerequired/not-traduttore',
+						'scm'        => 'git',
+						'is_private' => false,
 					],
-					'full_name'  => 'wearerequired/not-traduttore',
-					'scm'        => 'git',
-					'is_private' => false,
-				],
-			]
+				]
+			)
 		);
 		$signature = 'sha256=' . hash_hmac( 'sha256', $request->get_body(), 'traduttore-test' );
 		$request->add_header( 'x-event-key', 'repo:push' );
-		$request->add_header( 'x-hub-signature', $signature );
+		$request->add_header( 'x-hub-signature-256', $signature );
 		$response = rest_get_server()->dispatch( $request );
 
 		$this->assertErrorResponse( 404, $response );
@@ -110,27 +112,30 @@ class Bitbucket extends TestCase {
 
 	public function test_valid_project(): void {
 		$request = new WP_REST_Request( 'POST', '/traduttore/v1/incoming-webhook' );
-		$request->set_body_params(
-			[
-				'ref'        => 'refs/heads/master',
-				'repository' => [
-					'links'      => [
-						'html' => [
-							'href' => 'https://bitbucket.org/wearerequired/traduttore',
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body(
+			(string) wp_json_encode(
+				[
+					'ref'        => 'refs/heads/master',
+					'repository' => [
+						'links'      => [
+							'html' => [
+								'href' => 'https://bitbucket.org/wearerequired/traduttore',
+							],
 						],
+						'full_name'  => 'wearerequired/traduttore',
+						'scm'        => 'git',
+						'is_private' => false,
 					],
-					'full_name'  => 'wearerequired/traduttore',
-					'scm'        => 'git',
-					'is_private' => false,
-				],
-			]
+				]
+			)
 		);
 		$signature = 'sha256=' . hash_hmac( 'sha256', $request->get_body(), 'traduttore-test' );
 		$request->add_header( 'x-event-key', 'repo:push' );
-		$request->add_header( 'x-hub-signature', $signature );
+		$request->add_header( 'x-hub-signature-256', $signature );
 		$response = rest_get_server()->dispatch( $request );
 
-		$this->assertEquals( 200, $response->get_status() );
+		$this->assertSame( 200, $response->get_status() );
 		$this->assertSame( [ 'result' => 'OK' ], $response->get_data() );
 		$this->assertSame( Repository::VCS_TYPE_GIT, $this->project->get_repository_vcs_type() );
 		$this->assertSame( Repository::TYPE_BITBUCKET, $this->project->get_repository_type() );
@@ -145,27 +150,30 @@ class Bitbucket extends TestCase {
 		$this->project->set_repository_vcs_type( Repository::VCS_TYPE_HG );
 
 		$request = new WP_REST_Request( 'POST', '/traduttore/v1/incoming-webhook' );
-		$request->set_body_params(
-			[
-				'ref'        => 'refs/heads/master',
-				'repository' => [
-					'links'      => [
-						'html' => [
-							'href' => 'https://bitbucket.org/wearerequired/traduttore',
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body(
+			(string) wp_json_encode(
+				[
+					'ref'        => 'refs/heads/master',
+					'repository' => [
+						'links'      => [
+							'html' => [
+								'href' => 'https://bitbucket.org/wearerequired/traduttore',
+							],
 						],
+						'full_name'  => 'wearerequired/traduttore',
+						'scm'        => 'git',
+						'is_private' => false,
 					],
-					'full_name'  => 'wearerequired/traduttore',
-					'scm'        => 'git',
-					'is_private' => false,
-				],
-			]
+				]
+			)
 		);
 		$signature = 'sha256=' . hash_hmac( 'sha256', $request->get_body(), 'traduttore-test' );
 		$request->add_header( 'x-event-key', 'repo:push' );
-		$request->add_header( 'x-hub-signature', $signature );
+		$request->add_header( 'x-hub-signature-256', $signature );
 		$response = rest_get_server()->dispatch( $request );
 
-		$this->assertEquals( 200, $response->get_status() );
+		$this->assertSame( 200, $response->get_status() );
 		$this->assertSame( [ 'result' => 'OK' ], $response->get_data() );
 		$this->assertSame( Repository::VCS_TYPE_HG, $this->project->get_repository_vcs_type() );
 		$this->assertSame( Repository::TYPE_BITBUCKET, $this->project->get_repository_type() );
