@@ -1,8 +1,6 @@
 <?php
 /**
  * Class LegacyGitHub
- *
- * @package Traduttore\Tests
  */
 
 namespace Required\Traduttore\Tests\WebhookHandler;
@@ -10,22 +8,19 @@ namespace Required\Traduttore\Tests\WebhookHandler;
 use Required\Traduttore\Project;
 use Required\Traduttore\Repository;
 use Required\Traduttore\Tests\TestCase;
-use \WP_REST_Request;
+use WP_REST_Request;
 
 /**
  * Test cases for \Required\Traduttore\WebhookHandler\GitHub.
  */
 class LegacyGitHub extends TestCase {
-	/**
-	 * @var Project
-	 */
-	protected $project;
+	protected Project $project;
 
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$this->project = new Project(
-			$this->factory->project->create(
+			$this->factory()->project->create(
 				[
 					'name'                => 'Sample Project',
 					'source_url_template' => 'https://github.com/wearerequired/traduttore/blob/master/%file%#L%line%',
@@ -54,7 +49,7 @@ class LegacyGitHub extends TestCase {
 		$request->add_header( 'x-github-event', 'ping' );
 		$response = rest_get_server()->dispatch( $request );
 
-		$this->assertEquals( 200, $response->get_status() );
+		$this->assertSame( 200, $response->get_status() );
 		$this->assertSame( [ 'result' => 'OK' ], $response->get_data() );
 	}
 
@@ -68,10 +63,11 @@ class LegacyGitHub extends TestCase {
 
 	public function test_invalid_signature(): void {
 		$request = new WP_REST_Request( 'POST', '/github-webhook/v1/push-event' );
-		$request->set_body_params( [] );
-		$signature = 'sha1=' . hash_hmac( 'sha1', $request->get_body(), 'foo' );
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body( (string) wp_json_encode( [] ) );
+		$signature = 'sha256=' . hash_hmac( 'sha256', $request->get_body(), 'foo' );
 		$request->add_header( 'x-github-event', 'push' );
-		$request->add_header( 'x-hub-signature', $signature );
+		$request->add_header( 'x-hub-signature-256', $signature );
 		$response = rest_get_server()->dispatch( $request );
 
 		$this->assertErrorResponse( 'rest_forbidden', $response, 401 );
@@ -79,43 +75,49 @@ class LegacyGitHub extends TestCase {
 
 	public function test_invalid_branch(): void {
 		$request = new WP_REST_Request( 'POST', '/github-webhook/v1/push-event' );
-		$request->set_body_params(
-			[
-				'ref'        => 'refs/heads/master',
-				'repository' => [
-					'html_url'       => 'https://github.com/wearerequired/traduttore',
-					'default_branch' => 'develop',
-				],
-			]
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body(
+			(string) wp_json_encode(
+				[
+					'ref'        => 'refs/heads/master',
+					'repository' => [
+						'html_url'       => 'https://github.com/wearerequired/traduttore',
+						'default_branch' => 'develop',
+					],
+				]
+			)
 		);
-		$signature = 'sha1=' . hash_hmac( 'sha1', $request->get_body(), 'traduttore-test' );
+		$signature = 'sha256=' . hash_hmac( 'sha256', $request->get_body(), 'traduttore-test' );
 		$request->add_header( 'x-github-event', 'push' );
-		$request->add_header( 'x-hub-signature', $signature );
+		$request->add_header( 'x-hub-signature-256', $signature );
 		$response = rest_get_server()->dispatch( $request );
 
-		$this->assertEquals( 200, $response->get_status() );
+		$this->assertSame( 200, $response->get_status() );
 		$this->assertSame( [ 'result' => 'Not the default branch' ], $response->get_data() );
 	}
 
 	public function test_invalid_project(): void {
 		$request = new WP_REST_Request( 'POST', '/github-webhook/v1/push-event' );
-		$request->set_body_params(
-			[
-				'ref'        => 'refs/heads/master',
-				'repository' => [
-					'default_branch' => 'master',
-					'full_name'      => 'wearerequired/not-traduttore',
-					'html_url'       => 'https://github.com/wearerequired/not-traduttore',
-					'ssh_url'        => 'git@github.com:wearerequired/not-traduttore.git',
-					'clone_url'      => 'https://github.com/wearerequired/not-traduttore.git',
-					'url'            => 'https://github.com/wearerequired/not-traduttore',
-					'private'        => false,
-				],
-			]
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body(
+			(string) wp_json_encode(
+				[
+					'ref'        => 'refs/heads/master',
+					'repository' => [
+						'default_branch' => 'master',
+						'full_name'      => 'wearerequired/not-traduttore',
+						'html_url'       => 'https://github.com/wearerequired/not-traduttore',
+						'ssh_url'        => 'git@github.com:wearerequired/not-traduttore.git',
+						'clone_url'      => 'https://github.com/wearerequired/not-traduttore.git',
+						'url'            => 'https://github.com/wearerequired/not-traduttore',
+						'private'        => false,
+					],
+				]
+			)
 		);
-		$signature = 'sha1=' . hash_hmac( 'sha1', $request->get_body(), 'traduttore-test' );
+		$signature = 'sha256=' . hash_hmac( 'sha256', $request->get_body(), 'traduttore-test' );
 		$request->add_header( 'x-github-event', 'push' );
-		$request->add_header( 'x-hub-signature', $signature );
+		$request->add_header( 'x-hub-signature-256', $signature );
 		$response = rest_get_server()->dispatch( $request );
 
 		$this->assertErrorResponse( 404, $response );
@@ -123,26 +125,29 @@ class LegacyGitHub extends TestCase {
 
 	public function test_valid_project(): void {
 		$request = new WP_REST_Request( 'POST', '/github-webhook/v1/push-event' );
-		$request->set_body_params(
-			[
-				'ref'        => 'refs/heads/master',
-				'repository' => [
-					'full_name'      => 'wearerequired/traduttore',
-					'default_branch' => 'master',
-					'html_url'       => 'https://github.com/wearerequired/traduttore',
-					'ssh_url'        => 'git@github.com:wearerequired/traduttore.git',
-					'clone_url'      => 'https://github.com/wearerequired/traduttore.git',
-					'url'            => 'https://github.com/wearerequired/traduttore',
-					'private'        => false,
-				],
-			]
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body(
+			(string) wp_json_encode(
+				[
+					'ref'        => 'refs/heads/master',
+					'repository' => [
+						'full_name'      => 'wearerequired/traduttore',
+						'default_branch' => 'master',
+						'html_url'       => 'https://github.com/wearerequired/traduttore',
+						'ssh_url'        => 'git@github.com:wearerequired/traduttore.git',
+						'clone_url'      => 'https://github.com/wearerequired/traduttore.git',
+						'url'            => 'https://github.com/wearerequired/traduttore',
+						'private'        => false,
+					],
+				]
+			)
 		);
-		$signature = 'sha1=' . hash_hmac( 'sha1', $request->get_body(), 'traduttore-test' );
+		$signature = 'sha256=' . hash_hmac( 'sha256', $request->get_body(), 'traduttore-test' );
 		$request->add_header( 'x-github-event', 'push' );
-		$request->add_header( 'x-hub-signature', $signature );
+		$request->add_header( 'x-hub-signature-256', $signature );
 		$response = rest_get_server()->dispatch( $request );
 
-		$this->assertEquals( 200, $response->get_status() );
+		$this->assertSame( 200, $response->get_status() );
 		$this->assertSame( [ 'result' => 'OK' ], $response->get_data() );
 		$this->assertSame( Repository::VCS_TYPE_GIT, $this->project->get_repository_vcs_type() );
 		$this->assertSame( Repository::TYPE_GITHUB, $this->project->get_repository_type() );
@@ -159,26 +164,29 @@ class LegacyGitHub extends TestCase {
 		$this->project->set_repository_webhook_secret( $secret );
 
 		$request = new WP_REST_Request( 'POST', '/github-webhook/v1/push-event' );
-		$request->set_body_params(
-			[
-				'ref'        => 'refs/heads/master',
-				'repository' => [
-					'full_name'      => 'wearerequired/traduttore',
-					'default_branch' => 'master',
-					'html_url'       => 'https://github.com/wearerequired/traduttore',
-					'ssh_url'        => 'git@github.com:wearerequired/traduttore.git',
-					'clone_url'      => 'https://github.com/wearerequired/traduttore.git',
-					'url'            => 'https://github.com/wearerequired/traduttore',
-					'private'        => false,
-				],
-			]
+		$request->add_header( 'Content-Type', 'application/json' );
+		$request->set_body(
+			(string) wp_json_encode(
+				[
+					'ref'        => 'refs/heads/master',
+					'repository' => [
+						'full_name'      => 'wearerequired/traduttore',
+						'default_branch' => 'master',
+						'html_url'       => 'https://github.com/wearerequired/traduttore',
+						'ssh_url'        => 'git@github.com:wearerequired/traduttore.git',
+						'clone_url'      => 'https://github.com/wearerequired/traduttore.git',
+						'url'            => 'https://github.com/wearerequired/traduttore',
+						'private'        => false,
+					],
+				]
+			)
 		);
-		$signature = 'sha1=' . hash_hmac( 'sha1', $request->get_body(), $secret );
+		$signature = 'sha256=' . hash_hmac( 'sha256', $request->get_body(), $secret );
 		$request->add_header( 'x-github-event', 'push' );
-		$request->add_header( 'x-hub-signature', $signature );
+		$request->add_header( 'x-hub-signature-256', $signature );
 		$response = rest_get_server()->dispatch( $request );
 
-		$this->assertEquals( 200, $response->get_status() );
+		$this->assertSame( 200, $response->get_status() );
 		$this->assertSame( [ 'result' => 'OK' ], $response->get_data() );
 		$this->assertSame( Repository::VCS_TYPE_GIT, $this->project->get_repository_vcs_type() );
 		$this->assertSame( Repository::TYPE_GITHUB, $this->project->get_repository_type() );
