@@ -53,7 +53,6 @@ class Subversion extends Base {
 		);
 
 		return 0 === $status ? $target : null;
-
 	}
 
 	/**
@@ -70,7 +69,9 @@ class Subversion extends Base {
 		chdir( $target );
 		exec( escapeshellcmd( 'svn revert --recursive .' ), $output, $status );
 		exec( escapeshellcmd( 'svn update .' ), $output, $status );
-		chdir( $current_dir );
+		if ( $current_dir ) {
+			chdir( $current_dir );
+		}
 
 		return 0 === $status ? $target : null;
 	}
@@ -112,6 +113,6 @@ class Subversion extends Base {
 		 * @param string                          $checkout_url The URL to check out a Subversion repository.
 		 * @param \Required\Traduttore\Repository $repository   The current repository.
 		 */
-		return apply_filters( 'traduttore.svn_checkout_url', $checkout_url, $this->repository );
+		return apply_filters( 'traduttore.svn_checkout_url', (string) $checkout_url, $this->repository );
 	}
 }
