@@ -106,7 +106,7 @@ class Plugin {
 
 		add_action(
 			'traduttore.generate_zip',
-			static function( $translation_set_id ): void {
+			static function ( $translation_set_id ): void {
 				/** @var \GP_Translation_Set $translation_set */
 				$translation_set = GP::$translation_set->get( $translation_set_id );
 
@@ -117,7 +117,7 @@ class Plugin {
 
 		add_filter(
 			'gp_update_meta',
-			static function( $meta_tuple ) {
+			static function ( $meta_tuple ) {
 				$allowed_keys = [
 					Project::VERSION_KEY, // '_traduttore_version'.
 					Project::TEXT_DOMAIN_KEY, // '_traduttore_text_domain'.
@@ -183,14 +183,21 @@ class Plugin {
 
 		add_filter(
 			'slack_get_events',
-			static function( $events ) {
+			static function ( $events ) {
 				$events['traduttore.zip_generated'] = [
 					'action'      => 'traduttore.zip_generated',
 					'description' => __( 'When a new translation ZIP file is built', 'traduttore' ),
-					'message'     => function( $zip_path, $zip_url, GP_Translation_Set $translation_set ) {
+					'message'     => function ( $zip_path, $zip_url, GP_Translation_Set $translation_set ) {
 						/** @var \GP_Locale $locale */
-						$locale  = GP_Locales::by_slug( $translation_set->locale );
-						$project = new Project( GP::$project->get( $translation_set->project_id ) );
+						$locale = GP_Locales::by_slug( $translation_set->locale );
+
+						$gp_project = GP::$project->get( $translation_set->project_id );
+
+						if ( ! $gp_project ) {
+							return false;
+						}
+
+						$project = new Project( $gp_project );
 
 						/**
 						 * Filters whether a Slack notification for translation updates from GitHub should be sent.
@@ -231,7 +238,7 @@ class Plugin {
 				$events['traduttore.updated'] = [
 					'action'      => 'traduttore.updated',
 					'description' => __( 'When new translations are updated for a project', 'traduttore' ),
-					'message'     => function( Project $project, array $stats ) {
+					'message'     => function ( Project $project, array $stats ) {
 						[
 							$originals_added,
 							$originals_existing, // phpcs:ignore VariableAnalysis.CodeAnalysis.VariableAnalysis.UnusedVariable
@@ -392,6 +399,8 @@ class Plugin {
 	 *
 	 * @param \WP_REST_Request $request Request object.
 	 * @return bool True if permission is granted, false otherwise.
+	 *
+	 * @phpstan-param \WP_REST_Request<array{}> $request
 	 */
 	public function incoming_webhook_permission_callback( WP_REST_Request $request ): bool {
 		$result  = false;
@@ -422,8 +431,10 @@ class Plugin {
 	 *
 	 * @param \WP_REST_Request $request Request object.
 	 * @return \WP_Error|\WP_REST_Response REST response on success, error object on failure.
+	 *
+	 * @phpstan-param \WP_REST_Request<array{}> $request
 	 */
-	public function incoming_webhook_callback( WP_REST_Request $request ) {
+	public function incoming_webhook_callback( WP_REST_Request $request ): \WP_Error|\WP_REST_Response {
 		$result  = new \WP_Error( '400', 'Bad request' );
 		$handler = ( new WebhookHandlerFactory() )->get_handler( $request );
 
