@@ -1,38 +1,37 @@
 <?php
 /**
  * Class Configuration
- *
- * @package Traduttore\Tests
  */
 
 namespace Required\Traduttore\Tests;
 
-use \Required\Traduttore\Configuration as Config;
+use Required\Traduttore\Configuration as Config;
 
 /**
  * Test cases for \Required\Traduttore\Configuration.
  */
 class Configuration extends TestCase {
 	public function test_get_path(): void {
-		$config = new Config( dirname( __DIR__ ) . '/data/example-no-config' );
+		$config = new Config( \dirname( __DIR__ ) . '/data/example-no-config' );
 
-		$this->assertSame( dirname( __DIR__ ) . '/data/example-no-config', $config->get_path() );
+		$this->assertSame( \dirname( __DIR__ ) . '/data/example-no-config', $config->get_path() );
 	}
 
 	public function test_get_config_empty_directory(): void {
-		$config = new Config( dirname( __DIR__ ) . '/data/example-no-config' );
+		$config = new Config( \dirname( __DIR__ ) . '/data/example-no-config' );
 
 		$this->assertEmpty( $config->get_config() );
 	}
 
 	public function test_get_config_value_empty_directory(): void {
-		$config = new Config( dirname( __DIR__ ) . '/data/example-no-config' );
+		$config = new Config( \dirname( __DIR__ ) . '/data/example-no-config' );
 
+		// @phpstan-ignore-next-line
 		$this->assertNull( $config->get_config_value( 'foo' ) );
 	}
 
 	public function test_get_config_composer(): void {
-		$config = new Config( dirname( __DIR__ ) . '/data/example-with-composer' );
+		$config = new Config( \dirname( __DIR__ ) . '/data/example-with-composer' );
 
 		$this->assertEqualSets(
 			[
@@ -48,7 +47,7 @@ class Configuration extends TestCase {
 	}
 
 	public function test_get_config_traduttore(): void {
-		$config = new Config( dirname( __DIR__ ) . '/data/example-with-config' );
+		$config = new Config( \dirname( __DIR__ ) . '/data/example-with-config' );
 
 		$this->assertEqualSets(
 			[
