@@ -1,14 +1,12 @@
 <?php
 /**
  * Class ProjectLocator
- *
- * @package Traduttore\Tests
  */
 
 namespace Required\Traduttore\Tests;
 
-use \Required\Traduttore\Project;
-use \Required\Traduttore\ProjectLocator as Locator;
+use Required\Traduttore\Project;
+use Required\Traduttore\ProjectLocator as Locator;
 
 /**
  * Test cases for \Required\Traduttore\ProjectLocator.
@@ -29,21 +27,21 @@ class ProjectLocator extends TestCase {
 	 */
 	protected $subsub;
 
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
-		$this->root   = $this->factory->project->create(
+		$this->root   = $this->factory()->project->create(
 			[
 				'name' => 'Root',
 			]
 		);
-		$this->sub    = $this->factory->project->create(
+		$this->sub    = $this->factory()->project->create(
 			[
 				'name'              => 'Sub',
 				'parent_project_id' => $this->root->id,
 			]
 		);
-		$this->subsub = $this->factory->project->create(
+		$this->subsub = $this->factory()->project->create(
 			[
 				'name'                => 'SubSub',
 				'parent_project_id'   => $this->sub->id,
@@ -54,12 +52,6 @@ class ProjectLocator extends TestCase {
 
 	public function test_empty_string(): void {
 		$locator = new Locator( '' );
-
-		$this->assertNull( $locator->get_project() );
-	}
-
-	public function test_false(): void {
-		$locator = new Locator( false );
 
 		$this->assertNull( $locator->get_project() );
 	}
@@ -81,48 +73,55 @@ class ProjectLocator extends TestCase {
 		$project = $this->root;
 		$locator = new Locator( $project );
 
-		$this->assertEquals( $this->root->id, $locator->get_project()->get_id() );
+		$this->assertInstanceOf( Project::class, $locator->get_project() );
+		$this->assertSame( $this->root->id, $locator->get_project()->get_id() );
 	}
 
 	public function test_find_project_by_glotpress_path(): void {
 		$locator = new Locator( 'root' );
 
-		$this->assertEquals( $this->root->id, $locator->get_project()->get_id() );
+		$this->assertInstanceOf( Project::class, $locator->get_project() );
+		$this->assertSame( $this->root->id, $locator->get_project()->get_id() );
 	}
 
 	public function test_find_project_by_glotpress_subpath(): void {
 		$locator = new Locator( 'root/sub' );
 
-		$this->assertEquals( $this->sub->id, $locator->get_project()->get_id() );
+		$this->assertInstanceOf( Project::class, $locator->get_project() );
+		$this->assertSame( $this->sub->id, $locator->get_project()->get_id() );
 	}
 
 	public function test_find_project_by_glotpress_subsubpath(): void {
 		$locator = new Locator( 'root/sub/subsub' );
 
-		$this->assertEquals( $this->subsub->id, $locator->get_project()->get_id() );
+		$this->assertInstanceOf( Project::class, $locator->get_project() );
+		$this->assertSame( $this->subsub->id, $locator->get_project()->get_id() );
 	}
 
 	public function test_find_project_by_glotpress_id(): void {
 		$locator = new Locator( (int) $this->sub->id );
 
-		$this->assertEquals( $this->sub->id, $locator->get_project()->get_id() );
+		$this->assertInstanceOf( Project::class, $locator->get_project() );
+		$this->assertSame( $this->sub->id, $locator->get_project()->get_id() );
 	}
 
 	public function test_find_project_by_glotpress_id_as_string(): void {
 		$locator = new Locator( (string) $this->sub->id );
 
-		$this->assertEquals( $this->sub->id, $locator->get_project()->get_id() );
+		$this->assertInstanceOf( Project::class, $locator->get_project() );
+		$this->assertSame( $this->sub->id, $locator->get_project()->get_id() );
 	}
 
 	public function test_find_project_by_github_url(): void {
 		$locator = new Locator( 'https://github.com/wearerequired/traduttore' );
 
-		$this->assertEquals( $this->subsub->id, $locator->get_project()->get_id() );
+		$this->assertInstanceOf( Project::class, $locator->get_project() );
+		$this->assertSame( $this->subsub->id, $locator->get_project()->get_id() );
 	}
 
 	public function test_find_project_by_repository_name(): void {
 		$project = new Project(
-			$this->factory->project->create(
+			$this->factory()->project->create(
 				[
 					'name' => 'Foo Bar',
 				]
@@ -133,12 +132,13 @@ class ProjectLocator extends TestCase {
 
 		$locator = new Locator( 'wearerequired/traduttore-registry' );
 
-		$this->assertEquals( $project->get_id(), $locator->get_project()->get_id() );
+		$this->assertInstanceOf( Project::class, $locator->get_project() );
+		$this->assertSame( $project->get_id(), $locator->get_project()->get_id() );
 	}
 
 	public function test_find_project_by_repository_url(): void {
 		$project = new Project(
-			$this->factory->project->create(
+			$this->factory()->project->create(
 				[
 					'name' => 'Foo Bar',
 				]
@@ -149,6 +149,7 @@ class ProjectLocator extends TestCase {
 
 		$locator = new Locator( 'https://github.com/wearerequired/traduttore-registry' );
 
-		$this->assertEquals( $project->get_id(), $locator->get_project()->get_id() );
+		$this->assertInstanceOf( Project::class, $locator->get_project() );
+		$this->assertSame( $project->get_id(), $locator->get_project()->get_id() );
 	}
 }
